+++ documentation/example/inf/ecs/outputs.tf
@@ -1,0 +1,4 @@
+output "data" {
+  sensitive = true # Param data is always sensitive
+  value     = local.output_data_map
+}