+++ elasticache/subnet_group/variables.tf
@@ -1,0 +1,60 @@
+variable "group_map" {
+  type = map(object({
+    name_include_app_fields     = optional(bool)
+    name_infix                  = optional(bool)
+    vpc_az_key_list             = optional(list(string))
+    vpc_key                     = optional(string)
+    vpc_security_group_key_list = optional(list(string))
+    vpc_segment_key             = optional(string)
+  }))
+}
+
+variable "name_include_app_fields_default" {
+  type        = bool
+  default     = true
+  description = "If true, the Terraform project context will be included in the name"
+}
+
+variable "name_infix_default" {
+  type    = bool
+  default = true
+}
+
+variable "std_map" {
+  type = object({
+    name_replace_regex   = string
+    resource_name_prefix = string
+    resource_name_suffix = string
+    tags                 = map(string)
+  })
+}
+
+variable "vpc_az_key_list_default" {
+  type    = list(string)
+  default = ["a", "b"]
+}
+
+variable "vpc_data_map" {
+  type = map(object({
+    security_group_id_map = map(string)
+    segment_map = map(object({
+      route_public  = bool
+      subnet_id_map = map(string)
+    }))
+    vpc_cidr_block      = string
+    vpc_id              = string
+    vpc_ipv6_cidr_block = string
+  }))
+  default     = null
+  description = "Must be provided if one or more "
+}
+
+variable "vpc_key_default" {
+  type    = string
+  default = null
+}
+
+variable "vpc_segment_key_default" {
+  type    = string
+  default = "internal"
+}